requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pillow==10.3.0
numpy<2.0.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
typing-extensions==4.12.2
tensorflow-cpu
tf-keras
ml-dtypes
protobuf
google-generativeai
mysql-connector-python
requests